MGX_template/requirements.txt
streamlit
pytesseract
paddlepaddle
paddleocr
opencv-python
Pillow
pandas
numpy
plotly
altair
PyPDF2
python-docx
openpyxl